Rename package and file name to be consistent with the other services
