+++ api/config/cdr_config_local.json
@@ -183,7 +183,8 @@
       "accessTier": "controlled",
       "tanagraEnabled": true,
       "vwbTemplateId": "c843d45c-8030-4db8-9bf1-7a50545af6cf",
-      "publicReleaseNumber": 8
+      "publicReleaseNumber": 8,
+      "wgsCMRGVcfPath": "wgs/short_read/snpindel/cmrg"
     }
   ]
 }
