+++ api/src/main/java/org/pmiops/workbench/vwb/wsm/WsmClient.java
@@ -9,18 +9,18 @@ import java.util.UUID;
 import org.pmiops.workbench.config.WorkbenchConfig;
 import org.pmiops.workbench.exceptions.WorkbenchException;
 import org.pmiops.workbench.model.Workspace;
-import org.pmiops.workbench.wsmanager.ApiException;
-import org.pmiops.workbench.wsmanager.api.WorkspaceApi;
-import org.pmiops.workbench.wsmanager.model.CloneWorkspaceV2Request;
-import org.pmiops.workbench.wsmanager.model.CloneWorkspaceV2Result;
-import org.pmiops.workbench.wsmanager.model.GrantRoleRequestBody;
-import org.pmiops.workbench.wsmanager.model.IamRole;
-import org.pmiops.workbench.wsmanager.model.JobControl;
-import org.pmiops.workbench.wsmanager.model.JobReport;
-import org.pmiops.workbench.wsmanager.model.Properties;
-import org.pmiops.workbench.wsmanager.model.Property;
-import org.pmiops.workbench.wsmanager.model.State;
-import org.pmiops.workbench.wsmanager.model.WorkspaceDescription;
+import org.pmiops.workbench.vwb.wsmanager.ApiException;
+import org.pmiops.workbench.vwb.wsmanager.api.WorkspaceApi;
+import org.pmiops.workbench.vwb.wsmanager.model.CloneWorkspaceV2Request;
+import org.pmiops.workbench.vwb.wsmanager.model.CloneWorkspaceV2Result;
+import org.pmiops.workbench.vwb.wsmanager.model.GrantRoleRequestBody;
+import org.pmiops.workbench.vwb.wsmanager.model.IamRole;
+import org.pmiops.workbench.vwb.wsmanager.model.JobControl;
+import org.pmiops.workbench.vwb.wsmanager.model.JobReport;
+import org.pmiops.workbench.vwb.wsmanager.model.Properties;
+import org.pmiops.workbench.vwb.wsmanager.model.Property;
+import org.pmiops.workbench.vwb.wsmanager.model.State;
+import org.pmiops.workbench.vwb.wsmanager.model.WorkspaceDescription;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Qualifier;
