+++ api/src/main/java/org/pmiops/workbench/vwb/wsm/WsmConfig.java
@@ -6,10 +6,10 @@ import org.pmiops.workbench.auth.UserAuthentication;
 import org.pmiops.workbench.config.WorkbenchConfig;
 import org.pmiops.workbench.exceptions.ServerErrorException;
 import org.pmiops.workbench.vwb.common.VwbApiClientUtils;
-import org.pmiops.workbench.wsmanager.ApiClient;
-import org.pmiops.workbench.wsmanager.api.ControlledGcpResourceApi;
-import org.pmiops.workbench.wsmanager.api.ResourceApi;
-import org.pmiops.workbench.wsmanager.api.WorkspaceApi;
+import org.pmiops.workbench.vwb.wsmanager.ApiClient;
+import org.pmiops.workbench.vwb.wsmanager.api.ControlledGcpResourceApi;
+import org.pmiops.workbench.vwb.wsmanager.api.ResourceApi;
+import org.pmiops.workbench.vwb.wsmanager.api.WorkspaceApi;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
