+++ api/src/main/java/org/pmiops/workbench/workspaces/VwbWorkspaceServiceImpl.java
@@ -19,9 +19,9 @@ import org.pmiops.workbench.tanagra.model.FeatureSetList;
 import org.pmiops.workbench.utils.mappers.FirecloudMapper;
 import org.pmiops.workbench.utils.mappers.WorkspaceMapper;
 import org.pmiops.workbench.vwb.wsm.WsmClient;
-import org.pmiops.workbench.wsmanager.ApiException;
-import org.pmiops.workbench.wsmanager.model.IamRole;
-import org.pmiops.workbench.wsmanager.model.WorkspaceDescription;
+import org.pmiops.workbench.vwb.wsmanager.ApiException;
+import org.pmiops.workbench.vwb.wsmanager.model.IamRole;
+import org.pmiops.workbench.vwb.wsmanager.model.WorkspaceDescription;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
