+++ api/src/main/resources/vwb-ws-manager.yaml
